+++ src/main/java/com/tdtu/webproject/repository/CertificateRepositoryImp.java
@@ -3,7 +3,7 @@ package com.tdtu.webproject.repository;
 import com.tdtu.mbGenerator.generate.mybatis.example.TdtChungChiExample;
 import com.tdtu.mbGenerator.generate.mybatis.mapper.TdtChungChiMapper;
 import com.tdtu.mbGenerator.generate.mybatis.model.TdtChungChi;
-import com.tdtu.webproject.model.condition.CertificateCondition;
+import com.tdtu.webproject.mybatis.condition.CertificateCondition;
 import com.tdtu.webproject.utils.ArrayUtil;
 import com.tdtu.webproject.utils.DateUtil;
 import lombok.AllArgsConstructor;
