+++ src/main/java/com/tdtu/webproject/repository/LecturerRepository.java
@@ -1,9 +1,9 @@
 package com.tdtu.webproject.repository;
 
 import com.tdtu.mbGenerator.generate.mybatis.model.TdtGiangVien;
-import com.tdtu.webproject.model.condition.LecturerCondition;
-import com.tdtu.webproject.model.condition.LecturerTeachingHistoryCondition;
-import com.tdtu.webproject.model.condition.NormsLectureHoursCondition;
+import com.tdtu.webproject.mybatis.condition.LecturerCondition;
+import com.tdtu.webproject.mybatis.condition.LecturerTeachingHistoryCondition;
+import com.tdtu.webproject.mybatis.condition.NormsLectureHoursCondition;
 import com.tdtu.webproject.mybatis.result.*;
 
 import java.math.BigDecimal;
