+++ src/main/java/com/tdtu/webproject/repository/LecturerRepositoryImp.java
@@ -3,9 +3,9 @@ package com.tdtu.webproject.repository;
 import com.tdtu.mbGenerator.generate.mybatis.example.TdtGiangVienExample;
 import com.tdtu.mbGenerator.generate.mybatis.mapper.TdtGiangVienMapper;
 import com.tdtu.mbGenerator.generate.mybatis.model.TdtGiangVien;
-import com.tdtu.webproject.model.condition.LecturerCondition;
-import com.tdtu.webproject.model.condition.LecturerTeachingHistoryCondition;
-import com.tdtu.webproject.model.condition.NormsLectureHoursCondition;
+import com.tdtu.webproject.mybatis.condition.LecturerCondition;
+import com.tdtu.webproject.mybatis.condition.LecturerTeachingHistoryCondition;
+import com.tdtu.webproject.mybatis.condition.NormsLectureHoursCondition;
 import com.tdtu.webproject.mybatis.mapper.LecturerSupportMapper;
 import com.tdtu.webproject.mybatis.result.*;
 import com.tdtu.webproject.utils.ArrayUtil;
