+++ src/main/java/com/tdtu/webproject/repository/SubjectGroupRepositoryImp.java
@@ -3,7 +3,7 @@ package com.tdtu.webproject.repository;
 import com.tdtu.mbGenerator.generate.mybatis.example.TdtNhomMonExample;
 import com.tdtu.mbGenerator.generate.mybatis.mapper.TdtNhomMonMapper;
 import com.tdtu.mbGenerator.generate.mybatis.model.TdtNhomMon;
-import com.tdtu.webproject.model.condition.SubjectGroupCondition;
+import com.tdtu.webproject.mybatis.condition.SubjectGroupCondition;
 import com.tdtu.webproject.utils.ArrayUtil;
 import lombok.AllArgsConstructor;
 import org.springframework.context.annotation.ComponentScan;
