+++ src/main/java/com/tdtu/webproject/repository/TrainingProcessRepository.java
@@ -1,7 +1,7 @@
 package com.tdtu.webproject.repository;
 
 import com.tdtu.mbGenerator.generate.mybatis.model.TdtQuaTrinhDaoTao;
-import com.tdtu.webproject.model.condition.TrainingProcessCondition;
+import com.tdtu.webproject.mybatis.condition.TrainingProcessCondition;
 
 import java.math.BigDecimal;
 import java.util.List;
