+++ src/main/java/com/tdtu/webproject/service/ExtraLectureHoursService.java
@@ -1,6 +1,6 @@
 package com.tdtu.webproject.service;
 
-import com.tdtu.webproject.model.condition.NormsLectureHoursCondition;
+import com.tdtu.webproject.mybatis.condition.NormsLectureHoursCondition;
 import com.tdtu.webproject.mybatis.result.NormsLectureHoursResult;
 import com.tdtu.webproject.utils.ArrayUtil;
 import generater.openapi.model.ExtraHours;
