+++ src/main/java/com/tdtu/webproject/service/LecturerManageService.java
@@ -4,8 +4,8 @@ import com.tdtu.mbGenerator.generate.mybatis.model.TdtChungChi;
 import com.tdtu.mbGenerator.generate.mybatis.model.TdtDiemHaiLong;
 import com.tdtu.mbGenerator.generate.mybatis.model.TdtGiangVien;
 import com.tdtu.mbGenerator.generate.mybatis.model.TdtQuaTrinhDaoTao;
-import com.tdtu.webproject.model.condition.LecturerCondition;
-import com.tdtu.webproject.model.condition.NormsLectureHoursCondition;
+import com.tdtu.webproject.mybatis.condition.LecturerCondition;
+import com.tdtu.webproject.mybatis.condition.NormsLectureHoursCondition;
 import com.tdtu.webproject.mybatis.result.NormsLectureHoursResult;
 import com.tdtu.webproject.repository.LecturerRepository;
 import com.tdtu.webproject.utils.ArrayUtil;
