+++ src/main/java/com/tdtu/webproject/service/SatisfactionScoreService.java
@@ -2,7 +2,7 @@ package com.tdtu.webproject.service;
 
 import com.tdtu.mbGenerator.generate.mybatis.model.TdtDiemHaiLong;
 import com.tdtu.webproject.exception.BusinessException;
-import com.tdtu.webproject.model.condition.SatisfactionScoreCondition;
+import com.tdtu.webproject.mybatis.condition.SatisfactionScoreCondition;
 import com.tdtu.webproject.repository.SatisfactionScoreRepository;
 import com.tdtu.webproject.utils.ArrayUtil;
 import com.tdtu.webproject.utils.DateUtil;
