+++ src/main/java/com/tdtu/webproject/service/SubjectService.java
@@ -5,9 +5,9 @@ import com.tdtu.mbGenerator.generate.mybatis.model.TdtLoaiMon;
 import com.tdtu.mbGenerator.generate.mybatis.model.TdtMonHoc;
 import com.tdtu.mbGenerator.generate.mybatis.model.TdtNhomMon;
 import com.tdtu.webproject.exception.BusinessException;
-import com.tdtu.webproject.model.condition.SubjectCondition;
-import com.tdtu.webproject.model.condition.SubjectGroupCondition;
-import com.tdtu.webproject.model.condition.SubjectTrainingSysCondition;
+import com.tdtu.webproject.mybatis.condition.SubjectCondition;
+import com.tdtu.webproject.mybatis.condition.SubjectGroupCondition;
+import com.tdtu.webproject.mybatis.condition.SubjectTrainingSysCondition;
 import com.tdtu.webproject.repository.SubjectGroupRepository;
 import com.tdtu.webproject.repository.SubjectRepository;
 import com.tdtu.webproject.repository.SubjectTrainingSysRepository;
