+++ src/main/java/com/tdtu/webproject/service/TrainingProcessService.java
@@ -3,7 +3,7 @@ package com.tdtu.webproject.service;
 import com.tdtu.mbGenerator.generate.mybatis.model.TdtNgonNguDaoTao;
 import com.tdtu.mbGenerator.generate.mybatis.model.TdtQuaTrinhDaoTao;
 import com.tdtu.webproject.exception.BusinessException;
-import com.tdtu.webproject.model.condition.TrainingProcessCondition;
+import com.tdtu.webproject.mybatis.condition.TrainingProcessCondition;
 import com.tdtu.webproject.repository.TrainingLanguageRepository;
 import com.tdtu.webproject.repository.TrainingProcessRepository;
 import com.tdtu.webproject.utils.ArrayUtil;
